Places Remember/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
social-auth-app-django>=5.2.0
requests>=2.31.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
flake8>=6.1.0
